Various engine fixes for Muffin Mare
These are various fixes made while working on Muffin Mare. Some of these
fixes include:
  - Changing the PlayingSoundHandle type to be 32-bit from 16-bit
  - Added support to stop playing audio
  - Pass RenderingServices into `GameState::Resume`
  - Make StateMachine delete heap-allocated states if they have a flag
    enabled

+++ src/audio/PlayingSoundHandle.h
@@ -4,6 +4,6 @@
 #include <cstdint>
 
 namespace Adagio {
-    typedef std::uint16_t PlayingSoundHandle;
+    typedef std::uint32_t PlayingSoundHandle;
 }
 #endif //GL_ADAGIO_PLAYINGSOUNDHANDLE_H
